refactor!: rename top-level package to openfeature
