+++ openfeature/_backports/strenum.py
