+++ openfeature/provider/no_op_provider.py
@@ -1,11 +1,11 @@
 import typing
 
-from open_feature.evaluation_context import EvaluationContext
-from open_feature.flag_evaluation import FlagResolutionDetails, Reason
-from open_feature.hook import Hook
-from open_feature.provider.metadata import Metadata
-from open_feature.provider.no_op_metadata import NoOpMetadata
-from open_feature.provider.provider import AbstractProvider
+from openfeature.evaluation_context import EvaluationContext
+from openfeature.flag_evaluation import FlagResolutionDetails, Reason
+from openfeature.hook import Hook
+from openfeature.provider.metadata import Metadata
+from openfeature.provider.no_op_metadata import NoOpMetadata
+from openfeature.provider.provider import AbstractProvider
 
 PASSED_IN_DEFAULT = "Passed in default"
