+++ tests/test_flag_evaluation.py
@@ -1,5 +1,5 @@
-from open_feature.exception import ErrorCode
-from open_feature.flag_evaluation import FlagEvaluationDetails, Reason
+from openfeature.exception import ErrorCode
+from openfeature.flag_evaluation import FlagEvaluationDetails, Reason
 
 
 def test_evaulation_details_reason_should_be_a_string():
